ディスコネクト処理などを実装
